agents/requirements.txt
uagents
groq
pydantic>=2
python-dotenv
google-generativeai
pytest
pytest-asyncio
